Re-enable auto-user creation.
