+++ src/main/java/org/karmaexchange/provider/SocialNetworkProvider.java
@@ -1,6 +1,7 @@
 package org.karmaexchange.provider;
 
 import org.karmaexchange.dao.OAuthCredential;
+import org.karmaexchange.dao.User;
 
 public abstract class SocialNetworkProvider {
   protected OAuthCredential credential;
@@ -33,6 +34,8 @@ public abstract class SocialNetworkProvider {
   }
 
   public abstract boolean verifyCredential();
+
+  public abstract User createUser();
 
   public abstract String getProfileImageUrl();
 }
